+++ package.json
@@ -2,11 +2,8 @@
   "version": "1.0.0",
   "scripts": {
     "build-update-md-tables-from-openapi": "ts-node ./scripts/build-update-md-tables-from-openapi.ts",
-    "readme-fix-reference-docs": "ts-node ./scripts/readme-fix-reference-docs.ts",
     "remove-stoplight-tags-from-openapi": "ts-node ./scripts/remove-stoplight-tags-from-openapi.ts",
-    "remove-deprecated-from-openapi-remove-additional-properties": "ts-node ./scripts/remove-deprecated-from-openapi-remove-additional-properties.ts",
-    "generate-open-api-with-specific-endpoints": "ts-node ./scripts/generate-open-api-with-specific-endpoints.ts",
-    "generate-open-api-with-specific-endpoints-remove-additionalProperties-if": "ts-node ./scripts/generate-open-api-with-specific-endpoints-remove-additionalProperties-if.ts",
+    "prepare-open-api-for-sdk": "ts-node ./scripts/prepare-open-api-for-sdk/index.ts",
     "manage-project": "ts-node scripts/build-update-md-tables-from-openapi.ts && ts-node ./scripts/manage-project.ts",
     "readme-upload-missing-images": "ts-node ./scripts/readme-upload-missing-images.ts",
     "count-important-statistics-about-openapi": "ts-node ./scripts/count-important-statistics-about-openapi.ts"
