+++ scripts/bin/generate-open-api-with-specific-endpoints-remove-additionalProperties-if.ts
