+++ scripts/bin/generate-open-api-with-specific-endpoints.ts
