--- scripts/remove-deprecated-from-openapi-remove-additional-properties.ts
@@ -1,333 +1,0 @@
-import fsPromises from "fs/promises";
-import fs from "fs";
-import path from "path";
-import { omit } from "lodash";
-import minimist from "minimist";
-import colors from "colors";
-const options = minimist(process.argv.slice(2));
-
-function isObject(value) {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
-}
-
-const removeStoplightTag = (node: object): object => {
-  delete node["x-stoplight"];
-  for (const attr in node) {
-    if (isObject(node[attr])) {
-      removeStoplightTag(node[attr]);
-    }
-  }
-  return node;
-};
-
-const parseNullToNullable = (schemas) => {
-  let counter = 0;
-  const parseNullToNullableInSchema = (schema) => {
-    if (schema instanceof Object) {
-      if (schema.type === "null") {
-        counter++;
-        return { ...schema, nullable: true, type: "object" };
-      }
-
-      if (schema.properties) {
-        for (const propertyKey of Object.keys(schema.properties)) {
-          schema.properties[propertyKey] = parseNullToNullableInSchema(
-            schema.properties[propertyKey]
-          );
-        }
-      }
-    }
-
-    return schema;
-  };
-
-  for (const schemaName of Object.keys(schemas)) {
-    schemas[schemaName] = parseNullToNullableInSchema(schemas[schemaName]);
-  }
-
-  console.log(colors.green(`Replaced ${counter} nulls to nullable`));
-
-  return schemas;
-};
-
-const removeAdditionalProperties = (
-  e: any,
-  keepIfPropertiesNotPresent = false
-) => {
-  const simplifyObjectModel = (o) =>
-    omit(o, ["additionalProperties", "properties"]);
-
-  if (e instanceof Object) {
-    if (
-      !keepIfPropertiesNotPresent &&
-      "additionalProperties" in e &&
-      e.additionalProperties instanceof Object
-    ) {
-      return simplifyObjectModel(e);
-    } else if (
-      "additionalProperties" in e &&
-      e.additionalProperties instanceof Object &&
-      "properties" in e
-    ) {
-      return simplifyObjectModel(e);
-    }
-    for (const f of Object.keys(e)) {
-      if (typeof e[f] === "object") {
-        e[f] = removeAdditionalProperties(e[f]);
-      }
-    }
-  }
-  return e;
-};
-
-const skipList: { endpoint: string; methods: string[] | true }[] = [
-  { endpoint: "/v1/vouchers", methods: ["get", "post"] },
-  { endpoint: "/v1/vouchers/{code}", methods: ["put", "post"] },
-  { endpoint: "/v1/vouchers/import", methods: ["post"] },
-  { endpoint: "/v1/vouchers/qualification", methods: ["post"] },
-  { endpoint: "/v1/vouchers/bulk/async", methods: ["post"] },
-  { endpoint: "/v1/vouchers/metadata/async", methods: ["post"] },
-  { endpoint: "/v1/campaigns/qualification", methods: ["post"] },
-  { endpoint: "/v1/promotions/tiers", methods: ["get"] },
-  { endpoint: "/client/v1/promotions/tiers", methods: ["get"] },
-  { endpoint: "/v1/promotions/{campaignId}/tiers", methods: ["post"] },
-  { endpoint: "/v1/promotions/tiers/{promotionTierId}", methods: ["put"] },
-  { endpoint: "/v1/rewards", methods: ["get", "post"] },
-  { endpoint: "/v1/rewards/{rewardId}", methods: ["get", "put"] },
-  { endpoint: "/client/v1/validate", methods: ["get"] },
-  { endpoint: "/v1/promotions/validation", methods: ["post"] },
-  { endpoint: "/v1/promotions/tiers/{tierId}/validation", methods: ["post"] },
-  { endpoint: "/v1/vouchers/{code}/redemption", methods: ["post"] },
-  { endpoint: "/client/v1/redeem", methods: ["post"] },
-  {
-    endpoint: "/v1/promotions/tiers/{promotionTierId}/redemption",
-    methods: ["post"],
-  },
-  { endpoint: "/v1/loyalties", methods: ["get", "post"] },
-  { endpoint: "/v1/loyalties/{campaignId}", methods: ["get", "put"] },
-  { endpoint: "/v1/loyalties/{campaignId}/members", methods: ["get", "post"] },
-  {
-    endpoint: "/v1/loyalties/{campaignId}/members/{memberId}",
-    methods: ["get"],
-  },
-  { endpoint: "/v1/loyalties/members/{memberId}", methods: ["get"] },
-  {
-    endpoint: "/v1/loyalties/{campaignId}/members/{memberId}/activities",
-    methods: ["get"],
-  },
-  { endpoint: "/v1/loyalties/members/{memberId}/activities", methods: ["get"] },
-  {
-    endpoint: "/v1/loyalties/{campaignId}/points-expiration/export",
-    methods: ["post"],
-  },
-  {
-    endpoint: "/v1/loyalties/{campaignId}/earning-rules",
-    methods: ["get", "post"],
-  },
-  {
-    endpoint: "/v1/loyalties/{campaignId}/earning-rules/{earningRuleId}",
-    methods: ["put"],
-  },
-  {
-    endpoint: "/v1/loyalties/{campaignId}/reward-assignments",
-    methods: ["get"],
-  },
-  { endpoint: "/v1/loyalties/{campaignId}/rewards", methods: ["get", "post"] },
-  {
-    endpoint: "/v1/loyalties/{campaignId}/rewards/{assignmentId}",
-    methods: ["put"],
-  },
-  {
-    endpoint: "/v1/validation-rules/{validationRuleId}/assignments",
-    methods: ["post"],
-  },
-  { endpoint: "/v1/segments/{segmentId}", methods: ["get"] },
-  { endpoint: "/v1/segments", methods: ["post"] },
-  { endpoint: "/v1/consents", methods: ["get"] },
-  { endpoint: "/client/v1/consents#", methods: ["get"] },
-  { endpoint: "/v1/async-actions", methods: ["get"] },
-  { endpoint: "/v1/async-actions/{asyncActionId}", methods: ["get"] },
-  { endpoint: "/v1/metadata-schemas", methods: ["get"] },
-  { endpoint: "/v1/metadata-schemas/{resource}", methods: ["get"] },
-  { endpoint: "/v1/locations", methods: ["get"] },
-  { endpoint: "/v1/locations/{locationId}", methods: ["get"] },
-];
-
-///v1/loyalties/{campaignId}/members
-
-const main = async (keepIfPropertiesNotPresent) => {
-  const openApiPath = path.join(__dirname, "../reference/OpenAPI.json");
-  const openAPIContent = JSON.parse(
-    (await fsPromises.readFile(openApiPath)).toString()
-  );
-  removeStoplightTag(openAPIContent);
-
-  // Removing deprecated paths
-  const paths = {};
-  const pathsKeys = Object.keys(openAPIContent.paths);
-  for (const pathKey of pathsKeys) {
-    const skip = skipList.find((skip) => skip.endpoint === pathKey);
-    const path = {};
-    const methods = Object.keys(openAPIContent.paths[pathKey]);
-    for (const method of methods) {
-      if (
-        openAPIContent.paths[pathKey][method]?.deprecated ||
-        skip?.methods === true ||
-        (skip?.methods && skip.methods.includes(method))
-      ) {
-        continue;
-      }
-      path[method] = openAPIContent.paths[pathKey][method];
-      if (path[method].responses instanceof Object) {
-        path[method].responses = Object.fromEntries(
-          Object.entries(path[method].responses).filter((httpCodeAndSchema) => {
-            const [httpCode, schema] = httpCodeAndSchema;
-            return !isNaN(parseInt(httpCode)) && parseInt(httpCode) < 300;
-          })
-        );
-      }
-    }
-    if (Object.keys(path).length > 0) {
-      paths[pathKey] = path;
-    }
-  }
-
-  // Removing not used parameters
-  const parametersNames = JSON.stringify(paths)
-    .match(/"#\/components\/parameters\/.*?"/g)
-    .map((match) => match.replace('"#/components/parameters/', "").slice(0, -1))
-    .sort();
-
-  const parameters = {};
-  for (const parameterName of parametersNames) {
-    if (!openAPIContent.components.parameters?.[parameterName]) {
-      console.log(`not found ${parameterName} in parameters`);
-      continue;
-    }
-    parameters[parameterName] = removeAdditionalProperties(
-      openAPIContent.components.parameters[parameterName],
-      keepIfPropertiesNotPresent
-    );
-  }
-
-  // Removing not used schemas
-  const schemas = {};
-  const schemasNamesFoundInPaths = JSON.stringify(paths)
-    .match(/"#\/components\/schemas\/.*?"/g)
-    .map((match) => match.replace('"#/components/schemas/', "").slice(0, -1))
-    .sort();
-  // const schemasNamesFoundInParameters = JSON.stringify(
-  //   openAPIContent.components.parameters
-  // )
-  //   .match(/"#\/components\/schemas\/.*?"/g)
-  //   .map((match) => match.replace('"#/components/schemas/', "").slice(0, -1))
-  //   .sort();
-
-  const usedParameters = JSON.stringify(paths)
-    .match(/"#\/components\/parameters\/.*?"/g)
-    .map((match) =>
-      match.replace('"#/components/parameters/', "").slice(0, -1)
-    );
-  const schemasNamesFoundInPathsParameters: string[] = [];
-  for (const parameter of usedParameters) {
-    if (!openAPIContent.components.parameters?.[parameter]) {
-      continue;
-    }
-    const schemasNamesFoundInParameter = JSON.stringify(
-      openAPIContent.components.parameters[parameter]
-    )
-      .match(/"#\/components\/schemas\/.*?"/g)
-      .map((match) => match.replace('"#/components/schemas/', "").slice(0, -1));
-    schemasNamesFoundInPathsParameters.push(...schemasNamesFoundInParameter);
-  }
-
-  const allSchemasNames = [
-    ...schemasNamesFoundInPaths,
-    ...schemasNamesFoundInPathsParameters,
-  ];
-
-  for (const schemaName of allSchemasNames) {
-    if (!openAPIContent.components.schemas?.[schemaName]) {
-      console.log(`not found ${schemaName} in schemas`);
-      continue;
-    }
-    schemas[schemaName] = removeAdditionalProperties(
-      openAPIContent.components.schemas[schemaName],
-      keepIfPropertiesNotPresent
-    );
-  }
-
-  // Finding other schemas uses
-  let lastSchemaStringify = "";
-  while (true) {
-    const schemaStringify = JSON.stringify(schemas);
-    if (lastSchemaStringify === schemaStringify) {
-      break;
-    }
-    lastSchemaStringify = schemaStringify;
-    const schemasNames = schemaStringify
-      .match(/"#\/components\/schemas\/.*?"/g)
-      .map((match) => match.replace('"#/components/schemas/', "").slice(0, -1))
-      .sort();
-    for (const schemaName of schemasNames) {
-      if (!openAPIContent.components.schemas?.[schemaName]) {
-        console.log(`not found ${schemaName} in schemas`);
-        continue;
-      }
-      schemas[schemaName] = removeAdditionalProperties(
-        openAPIContent.components.schemas[schemaName],
-        keepIfPropertiesNotPresent
-      );
-    }
-  }
-
-  // //custom instructions
-  // if (
-  //   schemas?.["ExportResponseBase"]?.properties instanceof Object &&
-  //   Array.isArray(schemas?.["ExportResponseBase"]?.required)
-  // ) {
-  //   delete schemas["ExportResponseBase"].properties.result;
-  //   schemas["ExportResponseBase"].required = schemas[
-  //     "ExportResponseBase"
-  //   ].required.filter((fieldName) => fieldName !== "result");
-  // }
-  //
-
-  // Building all together
-  const newOpenApiFile = { ...openAPIContent };
-  newOpenApiFile.components.parameters = parameters;
-  // newOpenApiFile.components.schemas = schemas;
-  newOpenApiFile.components.schemas = parseNullToNullable(schemas);
-  newOpenApiFile.paths = paths;
-
-  //write the new OpenApiFile
-  const pathToTmp = path.join(__dirname, "../tmp");
-  if (!fs.existsSync(pathToTmp)) {
-    fs.mkdirSync(pathToTmp);
-  }
-  const pathToTmpReference = path.join(__dirname, "../tmp/reference");
-  if (!fs.existsSync(pathToTmpReference)) {
-    fs.mkdirSync(pathToTmpReference);
-  }
-  await fsPromises.writeFile(
-    path.join(__dirname, "../tmp/reference/OpenAPI.json"),
-    JSON.stringify(newOpenApiFile, null, 2)
-  );
-};
-
-const { always, usedWithStandardProperties } = options;
-if (!always && !usedWithStandardProperties) {
-  console.log(
-    colors.red(
-      "invalid arguments, missing `always` or `usedWithStandardProperties`"
-    )
-  );
-} else if (always && usedWithStandardProperties) {
-  console.log(
-    colors.red(
-      "invalid arguments, provided `always` and `usedWithStandardProperties`, please provide one argument"
-    )
-  );
-} else {
-  main(!!usedWithStandardProperties);
-}
