+++ scripts/types/globals.t.ts
